// packages/ml-ops/src/types.d.ts
export declare const MetricPointSchema: import("@sinclair/typebox").TObject<{
    step: import("@sinclair/typebox").TNumber;
    epoch: import("@sinclair/typebox").TNumber;
    trainLoss: import("@sinclair/typebox").TNumber;
    evalLoss: import("@sinclair/typebox").TOptional<import("@sinclair/typebox").TNumber>;
    learningRate: import("@sinclair/typebox").TNumber;
    gpuUtil: import("@sinclair/typebox").TNumber;
    gpuVramMB: import("@sinclair/typebox").TNumber;
    samplesPerSecond: import("@sinclair/typebox").TNumber;
    timestamp: import("@sinclair/typebox").TNumber;
}>;
export type MetricPoint = Static<typeof MetricPointSchema>;